pyproject.toml
[tool.poetry]
name = "django-hospital"
version = "0.1.0"
description = "Hospital and patient management API"
authors = ["Geonhyuk Im <athanasia9410@gmail.com>"]
readme = "README.md"
packages = [{include = "django_hospital"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
pytest = "^7.4.3"
pytest-django = "^4.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
